feat: 매장 수정 & 삭제 구현

+++ src/main/java/com/example/reservation/data/entity/Shop.java
@@ -25,7 +25,7 @@ public class Shop extends BaseEntity {
 
   private String shopExplain;
 
-  private double stars; // 총 별점
+  private int stars; // 총 별점
 
   private int reviewCount;
